pyproject.toml
[tool.poetry]
name = "clinical-cancer-test"
version = "0.1.0"
description = ""
authors = ["GitHubAlejandroDR <alejandrodominguezrecio@uma.es>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "1.4.1"
hydra-core = "^1.3.2"
scikit-learn = "^1.3.1"
pip = "^23.3"
ipykernel = "^6.25.2"
patsy = "^0.5.3"
hyperopt = "^0.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
